Relocated DogInfo and DogInfoDisplay into their own sub folders
